refactor: port 패키지명 in/out → internal/external 변경
- gitignore의 out/ 규칙으로 인해 port/out 하위 파일이 추적되지 않는 문제 해결
- port/in → port/internal (OrderUseCase)
- port/out → port/external (DepositPort, ProductPort)
- 관련 import 전체 수정 (service, kafka consumers, client adapters, controllers)
