+++ service/order/src/main/java/jabaclass/order/application/port/external/DepositPort.java
@@ -1,0 +1,9 @@
+package jabaclass.order.application.port.external;
+
+import java.util.UUID;
+import java.math.BigDecimal;
+
+public interface DepositPort {
+
+    boolean validateAndUse(UUID userId, BigDecimal depositAmount);
+}
