+++ service/order/src/main/java/jabaclass/order/application/port/internal/OrderUseCase.java
@@ -1,4 +1,4 @@
-package jabaclass.order.application.port.in;
+package jabaclass.order.application.port.internal;
 
 
 import java.util.List;
