+++ service/order/src/main/java/jabaclass/order/infrastructure/kafka/payment/PaymentExpiredEventConsumer.java
@@ -2,7 +2,7 @@ package jabaclass.order.infrastructure.kafka.payment;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import jabaclass.order.application.port.in.OrderUseCase;
+import jabaclass.order.application.port.internal.OrderUseCase;
 import jabaclass.order.infrastructure.kafka.payment.dto.PaymentExpiredEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
