+++ service/order/src/main/java/jabaclass/order/infrastructure/kafka/payment/PaymentFailedConsumer.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import jabaclass.order.application.port.in.OrderUseCase;
+import jabaclass.order.application.port.internal.OrderUseCase;
 import jabaclass.order.domain.model.PaymentResultStatus;
 import jabaclass.order.infrastructure.kafka.payment.dto.PaymentFailedEvent;
 import jabaclass.order.presentation.dto.request.UpdateOrderPaymentStatusRequestDto;
